+++ src/features/CategoryItem/CategoryItem.module.scss
@@ -1,0 +1,23 @@
+@use '../../shared/styles/utils/variables' as var;
+
+.category-list__item {
+  display: flex;
+  align-items: center;
+  min-height: 45px;
+  max-width: 270px;
+  background-color: var.$body-bg;
+  padding: 10px 40px 10px 14px;
+  margin-bottom: 5px;
+  border-radius: 5px;
+  color: var.$body-color;
+  cursor: pointer;
+}
+
+.category-list__item:hover {
+  color: var.$theme-primary-color;
+}
+
+.category-list__link {
+  text-decoration: none;
+  color: inherit;
+}
