+++ src/widgets/CategoryList/CategoryList.module.scss
@@ -3,9 +3,10 @@
 .category-list {
   padding: 30px;
   max-width: 330px;
-  max-height: calc(100vh - 300px);
+  height: fit-content;
   background-color: var.$white;
   border-radius: 10px;
+  width: inherit;
 }
 
 .category-list__title {
@@ -17,26 +18,4 @@ https://github.com/Studio-Yandex-Practicum/maxboom_frontend/issues/61 */
 .category-list__items {
   max-height: calc(100% - 50px);
   overflow-y: auto;
-}
-
-.category-list__item {
-  display: flex;
-  align-items: center;
-  min-height: 45px;
-  max-width: 270px;
-  background-color: var.$body-bg;
-  padding: 10px 40px 10px 14px;
-  margin-bottom: 5px;
-  border-radius: 5px;
-  color: var.$body-color;
-  cursor: pointer;
-}
-
-.category-list__item:hover {
-  color: var.$theme-primary-color;
-}
-
-.category-list__link {
-  text-decoration: none;
-  color: inherit;
 }
